rename : 워크스루 관련 파일 splash 폴더로 이동
