+++ src/components/splash/Splash.jsx
