+++ src/components/splash/Walkthrough.jsx
